requirements.txt
flask==2.2.2
flask-httpauth==4.7.0
flask-sqlalchemy==2.5.1
flask-restful==0.3.9
python-dateutil==2.8.2
requests==2.27.1